# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GitHub Manager API",
    "version": "1.0.0",
    "description": "API for managing GitHub repositories"
  },
  "servers": [
    {
      "url": "https://github-mcp-server-production-2326.up.railway.app"
    }
  ],
  "paths": {
    "/repos": {
      "get": {
        "operationId": "listRepositories",
        "summary": "List GitHub repositories",
        "parameters": [
          {
            "name": "visibility",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["all", "public", "private"],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of repositories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/file": {
      "get": {
        "operationId": "getFile",
        "summary": "Get file contents from a repository",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Repository owner username"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Repository name"
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "File path in the repository"
          },
          {
            "name": "branch",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Branch name (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "File content"
          }
        }
      },
      "post": {
        "operationId": "createOrUpdateFile",
        "summary": "Create or update a file in a repository",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "owner": { "type": "string", "description": "Repository owner" },
                  "repo": { "type": "string", "description": "Repository name" },
                  "path": { "type": "string", "description": "File path" },
                  "content": { "type": "string", "description": "File content" },
                  "message": { "type": "string", "description": "Commit message" },
                  "branch": { "type": "string", "description": "Branch name (optional)" }
                },
                "required": ["owner", "repo", "path", "content", "message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File created/updated"
          }
        }
      }
    },
    "/repos/{owner}/{repo}/branches": {
      "post": {
        "operationId": "createBranch",
        "summary": "Create a new branch",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branch": { "type": "string" },
                  "from_branch": { "type": "string" }
                },
                "required": ["branch"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch created"
          }
        }
      }
    },
    "/repos/{owner}/{repo}/pulls": {
      "post": {
        "operationId": "createPullRequest",
        "summary": "Create a pull request",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "body": { "type": "string" },
                  "head": { "type": "string" },
                  "base": { "type": "string" }
                },
                "required": ["title", "head", "base"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pull request created"
          }
        }
      }
    }
  }
}